requirements.txt
google-genai
selenium
undetected-chromedriver
beautifulsoup4
python-dotenv
ratelimit